@@ -4,11 +4,12 @@ import (
 	"context"
 	"encoding/base64"
 	"errors"
-	"github.com/google/go-github/v39/github"
-	"github.com/shurcooL/githubv4"
 	"regexp"
 	"strconv"
 	"time"
+
+	"github.com/google/go-github/v66/github"
+	"github.com/shurcooL/githubv4"
 )
 
 func (g *GitHubClient) listReleasesV4EnterPrice() ([]*github.RepositoryRelease, error) {
@@ -31,7 +32,7 @@ func (g *GitHubClient) listReleasesV4EnterPrice() ([]*github.RepositoryRelease, 
 		} `graphql:"repository(owner:$repositoryOwner,name:$repositoryName)"`
 	}
 
-	vars := map[string]interface{}{
+	vars := map[string]any{
 		"repositoryOwner": githubv4.String(g.owner),
 		"repositoryName":  githubv4.String(g.repository),
 		"releaseCursor":   (*githubv4.String)(nil),
@@ -102,7 +103,7 @@ func (g *GitHubClient) listReleasesV4() ([]*github.RepositoryRelease, error) {
 		} `graphql:"repository(owner:$repositoryOwner,name:$repositoryName)"`
 	}
 
-	vars := map[string]interface{}{
+	vars := map[string]any{
 		"repositoryOwner": githubv4.String(g.owner),
 		"repositoryName":  githubv4.String(g.repository),
 		"releaseCursor":   (*githubv4.String)(nil),
