@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	resource "github.com/concourse/github-release-resource"
-	"github.com/google/go-github/v39/github"
+	"github.com/google/go-github/v66/github"
 )
 
 type FakeGitHub struct {
